Update Jest tests
Removed allTimelines from reducer and redux so removing from tests.  Moved timeline tests into a different test file.
