+++ snprc_scheduler/src/client/tests/reducers/projectReducers.test.js
@@ -1,7 +1,6 @@
 import projectReducer from '../../reducers/projectReducer';
 import {projectData} from '../fixtures/projectData';
 import {animalData} from '../fixtures/animalData';
-import {timelineData} from "../fixtures/timelineData";
 
 import {
     ANIMAL_LIST_FILTERED,
@@ -11,10 +10,7 @@ import {
     PROJECT_LIST_RECEIVED,
     PROJECT_LIST_REQUEST_FAILED,
     PROJECT_LIST_SORTED,
-    PROJECT_SELECTED,
-    TIMELINE_DUPLICATED,
-    TIMELINE_LIST_RECEIVED,
-    TIMELINE_SELECTED
+    PROJECT_SELECTED
 } from "../../actions/dataActions";
 
 let projectState = {
@@ -34,13 +30,6 @@ const animalState = {
     animals: [...animalData]
 };
 
-const timelineState = {
-    0: timelineData[0],
-    1: timelineData[1],
-    errors: [],
-    allTimelines: [...timelineData],
-    timelines: [...timelineData]
-};
 test('make sure default state gets returns an empty error array.', () => {
     const state = projectReducer(undefined, {type: '@@init'});
     expect(state).toEqual({"errors": []});
@@ -152,48 +141,6 @@ test('Should select animal using filter (test id & gender)', () => {
                 animals: [animalData[1], animalData[2]]
             });
 
-});
-
-test('Should set timeline list', () => {
-    const action = {
-        type: TIMELINE_LIST_RECEIVED,
-        payload: timelineData
-    };
-    const state = projectReducer(timelineData, action);
-    expect(state).toEqual(
-            {
-                    ...timelineState,
-                    "selectedTimeline": {}
-
-            });
-});
-
-test('Should clone timeline returning a new timeline object with TimelineId = -1, revisionNum = -1, IsDraft = true', () => {
-    const action = {
-        type: TIMELINE_DUPLICATED,
-        payload: timelineData[0]
-    };
-    const state = projectReducer(timelineState, action);
-    expect(state).toEqual(
-            {
-                ...timelineState,
-                timelines: [ ...timelineData, {...timelineData[0], TimelineId: -1, revisionNum: -1, IsDraft: true} ]
-
-            });
-});
-
-
-test('Should select timeline', () => {
-    const action = {
-        type: TIMELINE_SELECTED,
-        payload: timelineData[0]
-    };
-    const state = projectReducer(timelineState, action);
-    expect(state).toEqual(
-            {
-                ...timelineState,
-                selectedTimeline: timelineData[0]
-            });
 });
 
 test('Should do a descending sort on project data', () => {
